+++ redux-toolkit/src/store.ts
@@ -1,9 +1,11 @@
-import { configureStore } from '@reduxjs/toolkit';
-import todoReducer from './Store/Slices/todoSlice'
+import { configureStore } from "@reduxjs/toolkit";
+import todoReducer from "./Store/Slices/todoSlice";
+
 export const store = configureStore({
   reducer: {
     todos: todoReducer,
   },
 });
-// export type RootState = ReturnType<typeof store.getState>
-// export type AppDispatch = typeof store.dispatch
+
+export type RootState = ReturnType<typeof store.getState>;
+//export type AppDispatch = typeof store.dispatch
